mols/requirements.txt
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=0.24.0
matplotlib>=3.4.0
rdkit>=2021.03.1
tqdm>=4.62.0 
gpytorch>=1.14
anthropic>=0.54.0
openai>=1.58.1
langchain>=0.1.0
langchain-experimental>=0.1.0
langchain-openai>=0.1.0
langchain-anthropic>=0.1.0
langgraph>=0.1.0
seaborn>=0.11.0
overrides>=7.0.0
ipykernel>=6.27.1
jupyter>=1.0.0
notebook>=7.4.4